+++ ts/features/itwallet/common/store/actions/index.ts
@@ -1,10 +1,10 @@
 import { ItwCredentialsActions } from "../../../credentials/store/actions";
-import { ItwIdentificationActions } from "../../../identification/store/actions";
+import { ItwIdentificationActions } from "../../../identification/common/store/actions";
 import { ItwIssuanceActions } from "../../../issuance/store/actions";
 import { ItwLifecycleActions } from "../../../lifecycle/store/actions";
 import { ItwWalletInstanceActions } from "../../../walletInstance/store/actions";
-import { ItwPreferencesActions } from "./preferences";
 import { ItwEnvironmentActions } from "./environment";
+import { ItwPreferencesActions } from "./preferences";
 import { ItwSecurePreferencesActions } from "./securePreferences";
 
 export type ItwActions =
