+++ ts/features/itwallet/identification/cie/saga/__tests__/checkHasNfcFeatureSaga.test.ts
@@ -1,0 +1,19 @@
+import cieManager from "@pagopa/react-native-cie";
+import { testSaga } from "redux-saga-test-plan";
+import { checkHasNfcFeatureSaga } from "..";
+import { itwHasNfcFeature } from "../../../common/store/actions";
+
+describe("checkHasNfcFeatureSaga", () => {
+  test.each([true, false])(
+    "If hasNFCFeature returns %p, should update the state accordingly",
+    arg => {
+      testSaga(checkHasNfcFeatureSaga)
+        .next()
+        .call(cieManager.hasNFCFeature)
+        .next(arg)
+        .put(itwHasNfcFeature.success(arg))
+        .next()
+        .isDone();
+    }
+  );
+});
