+++ ts/features/itwallet/identification/cie/screens/ItwActivateNfcScreen.tsx
@@ -7,15 +7,15 @@ import {
 import { useFocusEffect } from "@react-navigation/native";
 import { useCallback } from "react";
 import { Alert } from "react-native";
-import { IOScrollViewWithLargeHeader } from "../../../../components/ui/IOScrollViewWithLargeHeader";
-import I18n from "../../../../i18n";
-import * as cieUtils from "../../../authentication/login/cie/utils/cie";
+import { IOScrollViewWithLargeHeader } from "../../../../../components/ui/IOScrollViewWithLargeHeader";
+import I18n from "../../../../../i18n";
+import * as cieUtils from "../../../../authentication/login/cie/utils/cie";
 import {
   trackItWalletCieNfcActivation,
   trackItWalletCieNfcGoToSettings
-} from "../../analytics";
-import { ItwEidIssuanceMachineContext } from "../../machine/eid/provider";
-import { isL3FeaturesEnabledSelector } from "../../machine/eid/selectors";
+} from "../../../analytics";
+import { ItwEidIssuanceMachineContext } from "../../../machine/eid/provider";
+import { isL3FeaturesEnabledSelector } from "../../../machine/eid/selectors";
 
 export const ItwActivateNfcScreen = () => {
   const machineRef = ItwEidIssuanceMachineContext.useActorRef();
