+++ ts/features/itwallet/identification/cieId/hooks/useCieIdApp.ts
@@ -9,10 +9,10 @@ import {
   CIE_ID_ERROR_MESSAGE,
   IO_LOGIN_CIE_SOURCE_APP,
   IO_LOGIN_CIE_URL_SCHEME
-} from "../../../authentication/login/cie/utils/cie";
-import { isAndroid, isIos } from "../../../../utils/platform";
-import { convertUnknownToError } from "../../../../utils/errors";
-import { ItwEidIssuanceMachineContext } from "../../machine/eid/provider";
+} from "../../../../authentication/login/cie/utils/cie";
+import { isAndroid, isIos } from "../../../../../utils/platform";
+import { convertUnknownToError } from "../../../../../utils/errors";
+import { ItwEidIssuanceMachineContext } from "../../../machine/eid/provider";
 
 type CieIdHookResult = {
   /**
