+++ ts/features/itwallet/identification/common/store/selectors/index.ts
@@ -1,5 +1,5 @@
 import * as pot from "@pagopa/ts-commons/lib/pot";
-import { GlobalState } from "../../../../../store/reducers/types";
+import { GlobalState } from "../../../../../../store/reducers/types";
 
 export const itwHasNfcFeatureSelector = (state: GlobalState) =>
   pot.getOrElse(state.features.itWallet.identification.hasNfcFeature, false);
