+++ ts/features/itwallet/identification/spid/screens/ItwIdentificationIdpSelectionScreen.tsx
@@ -1,19 +1,19 @@
 import { VSpacer } from "@pagopa/io-app-design-system";
 import { useFocusEffect } from "@react-navigation/native";
 import { useRef, useCallback } from "react";
-import { isReady } from "../../../../common/model/RemoteValue";
-import IdpsGrid from "../../../authentication/login/idp/components/IdpsGrid";
-import { IOScrollViewWithLargeHeader } from "../../../../components/ui/IOScrollViewWithLargeHeader";
-import { randomOrderIdps } from "../../../authentication/login/idp/screens/IdpSelectionScreen";
-import { loadIdps } from "../../../../store/actions/content";
-import { useIODispatch, useIOSelector } from "../../../../store/hooks";
-import { idpsRemoteValueSelector } from "../../../../store/reducers/content";
-import { idps as idpsFallback, SpidIdp } from "../../../../utils/idps";
-import { ItwEidIssuanceMachineContext } from "../../machine/eid/provider";
+import { isReady } from "../../../../../common/model/RemoteValue";
+import IdpsGrid from "../../../../authentication/login/idp/components/IdpsGrid";
+import { IOScrollViewWithLargeHeader } from "../../../../../components/ui/IOScrollViewWithLargeHeader";
+import { randomOrderIdps } from "../../../../authentication/login/idp/screens/IdpSelectionScreen";
+import { loadIdps } from "../../../../../store/actions/content";
+import { useIODispatch, useIOSelector } from "../../../../../store/hooks";
+import { idpsRemoteValueSelector } from "../../../../../store/reducers/content";
+import { idps as idpsFallback, SpidIdp } from "../../../../../utils/idps";
+import { ItwEidIssuanceMachineContext } from "../../../machine/eid/provider";
 import {
   trackItWalletSpidIDPSelected,
   trackItWalletSpidIDPSelection
-} from "../../analytics";
+} from "../../../analytics";
 
 export const ItwIdentificationIdpSelectionScreen = () => {
   const dispatch = useIODispatch();
